requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
drf-yasg